Refactoring: separate HydraAuctionUtils.Tx.Build
